fix: likecomment

+++ src/like-comment/like-comment.service.ts
@@ -6,6 +6,7 @@ import { LikeComment } from './entities/like-comment.entity';
 import { Repository } from 'typeorm';
 import { USER_ID_HEADER_NAME } from 'src/auth/constant';
 import { error } from 'console';
+import { Exception } from 'handlebars';
 
 @Injectable()
 export class LikeCommentService {
@@ -14,22 +15,22 @@ export class LikeCommentService {
     private readonly likePostRepository: Repository<LikeComment>
   ) { }
 
-  async create(createLikeCommentDto: CreateLikeCommentDto,req:Request) {
-    const creater= req.headers[USER_ID_HEADER_NAME]
-    createLikeCommentDto['user']=creater
-    try{
+  async create(createLikeCommentDto: CreateLikeCommentDto, req: Request) {
+    const creater = req.headers[USER_ID_HEADER_NAME]
+    createLikeCommentDto['user'] = creater
+    try {
       await this.likePostRepository.save(createLikeCommentDto);
       return {
-        status:1,
-        message:"OK"
+        status: 1,
+        message: "OK"
       }
-    }catch(error){
+    } catch (error) {
       return {
-        status:-1,
-        message:"Tè"+error
+        status: -1,
+        message: "Tè" + error
       }
     }
-    
+
   }
 
   async findAll(): Promise<LikeComment[]> {
@@ -57,22 +58,26 @@ export class LikeCommentService {
     return await this.likePostRepository
       .createQueryBuilder('l')
       .leftJoin('l.user', 'user')
-      .addSelect(['user.fullname','user.id','user.avatar'])
+      .addSelect(['user.fullname', 'user.id', 'user.avatar'])
       .where({
         comment: comment_id
       })
       .orderBy({
-        'l.create_at':'DESC'
+        'l.create_at': 'DESC'
       })
       .getMany()
 
   }
 
-  async update(user_id: number,comment_id:number, updateLikeCommentDto: UpdateLikeCommentDto) {
+  async update(user_id: number, comment_id: number, updateLikeCommentDto: UpdateLikeCommentDto) {
+    if (!comment_id || !user_id) {
+      throw Exception
+    }
+
     const likeOld = await this.likePostRepository.findOne({
       where: {
         user: user_id,
-        comment:comment_id
+        comment: comment_id
       }
     })
     const likeUpdate = await this.likePostRepository.save({
@@ -82,23 +87,25 @@ export class LikeCommentService {
     return likeUpdate;
   }
 
-  async remove(comment_id: number,user_id:number) {
+  async remove(comment_id: number, user_id: number) {
     try {
-
+      if (!comment_id || !user_id) {
+        throw Exception
+      }
       await this.likePostRepository.delete({
-        comment:comment_id,
-        user:user_id
+        comment: comment_id,
+        user: user_id
       })
 
       return {
-        message:"Delete Success",
-        status:"OK"
+        message: "Delete Success",
+        status: "OK"
       };
 
     } catch (error) {
       return {
-        message:"Delete Failed",
-        status:error
+        message: "Delete Failed",
+        status: error
       };
 
     }
